refactor: extract CORS configuration to Hive.Abstractions
Moved CORS configuration models and validators from Hive.MicroServices to
Hive.Abstractions under the Hive.Configuration.CORS namespace to enable
reuse across different hosting models (MicroServices, Functions, etc.).

Changes:
- Moved Options.cs, CORSPolicy.cs, CORSPolicyValidator.cs, and OptionsValidator.cs
  to hive.core/src/Hive.Abstractions/Configuration/CORS/
- Updated namespace from Hive.MicroServices.CORS to Hive.Configuration.CORS
- Removed ASP.NET Core dependency from CORSPolicy (ToCORSPolicyBuilderAction method)
- Inlined policy builder logic in Extension.cs as BuildCorsPolicy() helper method
- Updated test files to use new namespace

All tests passing:
- Unit tests: 25/25 passed
- Integration tests: 18/18 passed
- Full build: 225/225 tests passed

This follows Option A from CORS_EXTRACTION_ANALYSIS.md - moving shared
abstractions to Hive.Abstractions while keeping ASP.NET Core specific
middleware in Hive.MicroServices.
