+++ hive.core/src/Hive.Abstractions/Configuration/CORS/OptionsValidator.cs
@@ -1,7 +1,7 @@
 using FluentValidation;
 using Hive.Extensions;
 
-namespace Hive.MicroServices.CORS;
+namespace Hive.Configuration.CORS;
 
 /// <summary>
 /// CORS OptionsValidator based on FluentValidation
